tests: move command tests
